+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Updated behavior with unresolved inherited dependencies (#633)
 - Updated behavior to only update installed dependencies on `pilet upgrade`
+- Added check to prevent installing of invalid versions locally (#635)
 
 ## 1.3.0 (October 9, 2023)